+++ test/spec/BpmnPropertiesPanelRenderer.spec.js
@@ -36,6 +36,7 @@ import CloudElementTemplatesPropertiesProvider from 'src/provider/cloud-element-
 
 import ElementTemplateChooserModule from '@bpmn-io/element-template-chooser';
 import ElementTemplatesIconsRenderer from '@bpmn-io/element-templates-icons-renderer';
+import ConnectorsExtensionModule from 'bpmn-js-connectors-extension';
 
 import CamundaBehaviorsModule from 'camunda-bpmn-js-behaviors/lib/camunda-platform';
 import ZeebeBehaviorsModule from 'camunda-bpmn-js-behaviors/lib/camunda-cloud';
@@ -248,7 +249,8 @@ describe('<BpmnPropertiesPanelRenderer>', function() {
           BpmnPropertiesProvider,
           CloudElementTemplatesPropertiesProvider,
           ElementTemplateChooserModule,
-          ElementTemplatesIconsRenderer
+          ElementTemplatesIconsRenderer,
+          ConnectorsExtensionModule
         ],
         moddleExtensions: {
           zeebe: ZeebeModdle